+++ src/headless/HeadlessUtils.h
@@ -11,7 +11,7 @@ namespace Surge
 namespace Headless
 {
 
-SurgeSynthesizer* createSurge(int sr);
+std::shared_ptr<SurgeSynthesizer> createSurge(int sr);
 
 void writeToStream(const float* data, int nSamples, int nChannels, std::ostream& str);
 void writeToWav(const float* data, int nSamples, int nChannels, float sampleRate, std::string wavFileName);
